Add Media ordering to All formatters that need it
Which are all. I mean, we need to. This is using @giancarlobi
