+++ src/Plugin/Field/FieldFormatter/Strawberry3DFormatter.php
@@ -167,6 +167,9 @@ class Strawberry3DFormatter extends StrawberryBaseFormatter {
       }*/
       $i = 0;
       if (isset($jsondata[$key])) {
+        // Order 3D Models based on a given 'sequence' key
+        $ordersubkey = 'sequence';
+        $this->orderSequence($jsondata, $key, $ordersubkey);
         foreach ($jsondata[$key] as $mediaitem) {
           $i++;
           if ($i > $number_models) {
