+++ src/Plugin/Field/FieldFormatter/StrawberryMediaFormatter.php
@@ -172,6 +172,9 @@ class StrawberryMediaFormatter extends StrawberryBaseFormatter {
       $elements[$delta]['#attached']['library'][] = 'format_strawberryfield/iiif_openseadragon_strawberry';
 
       if (isset($jsondata[$key])) {
+        // Order Images based on a given 'sequence' key
+        $ordersubkey = 'sequence';
+        $this->orderSequence($jsondata, $key, $ordersubkey);
         $iiifhelper = new IiifHelper($this->getIiifUrls()['public'], $this->getIiifUrls()['internal']);
         foreach ($jsondata[$key] as $mediaitem) {
           $i++;
